refactor(secutiry): переименовать модуль с моделями
